Rename input components
