+++ frontend/src/lib/forms/index.ts
@@ -3,7 +3,8 @@ import SubmitButton from './SubmitButton.svelte';
 import Form from './Form.svelte';
 import FormField from './FormField.svelte';
 import FormError from './FormError.svelte';
-import Input from './InputFormField.svelte';
+import Input from './Input.svelte';
+import PlainInput from './PlainInput.svelte';
 import Checkbox from './Checkbox.svelte';
 import ProtectedForm, { type Token } from './ProtectedForm.svelte';
 import Select from './Select.svelte';
@@ -23,6 +24,7 @@ export {
   FormField,
   FormError,
   Input,
+  PlainInput,
   ProtectedForm,
   Select,
   TextArea,
